LayoutTests/imported/w3c/web-platform-tests/speculation-rules/prefetch/redirect-middle-of-prefetch.https_origin=same-origin-expected.txt


PASS Navigate to the final URL of the prefetch
PASS Navigation is redirected to the final URL of the prefetch
